src/requirements.txt
paho-mqtt
pulsar-client